feat: Move ETH form to separate component
