+++ src/libs/components/components.ts
@@ -1,2 +1,3 @@
-export { Loader } from "./loader/loader.tsx";
 export { Error } from "./error/error.tsx";
+export { Loader } from "./loader/loader.tsx";
+export { EthUsdtForm } from "./eth-usdt-form/eth-usdt-form.tsx";
