--- src/libs/components/error/styles.css
@@ -1,9 +1,0 @@
-.error {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  color: hsl(0 100% 50%);
-  font-size: 30px;
-  font-weight: bold;
-}
