+++ src/libs/components/eth-usdt-form/styles.module.css
@@ -1,0 +1,74 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  width: 300px;
+  max-width: 100%;
+  padding: 20px;
+  border: 1px solid hsl(0 4% 76%);
+  border-radius: 15px;
+}
+
+.label {
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+
+.label-text {
+  font-size: 18px;
+}
+
+.amount-input,
+.receive-input {
+  height: 30px;
+  font-size: 16px;
+  padding: 5px;
+  border: 1px solid hsl(0 4% 76%);
+  border-radius: 8px;
+}
+
+:is(.amount-input, .receive-input):focus {
+  border: 1px solid hsl(0 2% 52%);
+  background: hsl(0 17% 94%);
+}
+
+.checkbox-wrapper {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.checkbox {
+  appearance: none;
+  position: relative;
+  display: flex;
+  align-items: center;
+  width: 60px;
+  height: 25px;
+  border: 1px solid hsl(0 4% 76%);
+  border-radius: 15px;
+  cursor: pointer;
+}
+
+.checkbox::after {
+  content: "";
+  position: absolute;
+  left: 2px;
+  width: 20px;
+  height: 20px;
+  background: hsl(0 17% 94%);
+  border: 1px solid hsl(0 2% 52%);
+  border-radius: 50%;
+  transition: 0.3s ease;
+}
+
+.checkbox.active::after {
+  left: calc(100% - 2px);
+  transform: translateX(-100%);
+}
+
+.error {
+  color: hsl(0 100% 50%);
+  font-size: 14px;
+}
