+++ src/libs/components/loader/loader.tsx
@@ -1,7 +1,7 @@
-import "./styles.css";
+import styles from "./styles.module.css";
 
 function Loader() {
-  return <div className="loader">Loading...</div>;
+  return <div className={styles.loader}>Loading...</div>;
 }
 
 export { Loader };
