--- src/libs/components/loader/styles.css
@@ -1,9 +1,0 @@
-.loader {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  color: hsl(120 35% 51%);
-  font-size: 30px;
-  font-weight: bold;
-}
