+++ src/libs/components/loader/styles.module.css
@@ -1,0 +1,5 @@
+.loader {
+  color: hsl(120 35% 51%);
+  font-size: 30px;
+  font-weight: bold;
+}
